fixes #5
The issue was that the stack is allocated a variable amount of space upon launching the process. At times, overflows completely overwrite the stack, which causes the corefile object to abort parsing the stack. 

To resolve this, a new smash_stack() function was made which will check to see if the corefile's stack is overwritten. If so, the function does a recursive call to launch a new process until enough space on the stack is allocated. This issue arises infrequently, so there's no significant performance loss.